Finished Controller and made sure links were working properly
