+++ LacamasFair/Views/Facility/FacilityRental.cshtml
@@ -17,7 +17,7 @@
 @if (isAdmin == true)
 {
     <div class="text-center">
-        <a class="btn btn-primary text-dark" asp-action="AddFacilityRental" asp-controller="Home">Add a Facility Rental</a>
+        <a class="btn btn-primary text-dark" asp-action="AddFacilityRental" asp-controller="Facility">Add a Facility Rental</a>
     </div>
 }
