+++ LacamasFair/Views/Shared/_Layout.cshtml
@@ -19,7 +19,7 @@
                                 <a class="nav-link text-white" asp-area="" asp-controller="Home" asp-action="HistoryAndGoal">About Us</a>
                             </li>
                             <li class="nav-item">
-                                <a class="nav-link text-white" asp-area="" asp-controller="Home" asp-action="FacilityRental">Facility Rental</a>
+                                <a class="nav-link text-white" asp-area="" asp-controller="Facility" asp-action="FacilityRental">Facility Rental</a>
                             </li>
                             <li class="nav-item">
                                 <a class="nav-link text-white" asp-area="" asp-controller="Home" asp-action="CommunityCenter">Community Center</a>
